FEAT:Added streaming and pivoted to flask
